polybar-configs
omit battery-arch, popup-calendar if not using a battery and if not interested in using the popup-calendar. If you wanted to use the battery-arch script, just follow the first part in installing the battery custom script in the polybar repo. As for the pop-up calendar, i used this script from the polybar repository: https://github.com/polybar/polybar-scripts/tree/master/polybar-scripts/popup-calendar